Adding a fallback to vconftool
This patch allows the creation of images when buxton is used.

During the process of creating the images, the serveur buxton
is not started. Unfortunately, the client library of buxton
dont allow a fall back to the direct access of the databases.

This patch fallback to use the command vconftool that works
even when the server buxton is off.

See https://github.com/sofar/buxton/issues/144

Change-Id: I89e68deda827ef79670250134badcaf2106d2223
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>